refactor(sync): extract skills read/write to claude/skills.rs
Following the plugin_assets.rs pattern (FU-2 of v0.8.0 refinement
follow-up). Move `read_skills` (162 LOC) and `write_skills` (63 LOC)
out of claude/mod.rs into claude/skills.rs as free functions; the
trait methods become single-line delegations.

Also bump the inherent dir-helper accessors (commands_dir,
skills_dir, hooks_dir, agents_dir, settings_path,
instructions_path) from `fn` to `pub(super) fn` so subsequent
topic extracts can derive paths without re-implementing them.

claude/mod.rs:    1853 → 1631 LOC
claude/skills.rs: new (252 LOC including module docstring)

312 sync-lib tests pass.
